Fixed reviewed code for the searchbar feature
